Finish OOP
