+++ src/main/java/ExitCommand.java
@@ -1,0 +1,13 @@
+public class ExitCommand extends Command{
+    public ExitCommand(String desc) {
+        super(desc);
+    }
+
+    public boolean isExit() {
+        return true;
+    }
+
+    public void execute(TaskList tasks, Storage storage) {
+
+    }
+}
